--- resolver/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "fluent-reference-resolver",
-  "private": true,
-  "version": "0.0.1",
-  "description": "A reference resolver for Fluent AST",
-  "scripts": {
-    "watch": "tsc --watch",
-    "test": "mocha -u tdd -r ts-node/register test/**/*_test.ts"
-  },
-  "author": "Mozilla <l10n-drivers@mozilla.org>",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@types/minimist": "^1.2.0",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.0.4",
-    "mocha": "^6.1.4",
-    "ts-node": "^8.2.0",
-    "typescript": "^3.5.1"
-  }
-}
