--- resolver/test/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2018",
-    "module": "commonjs",
-    "strict": true,
-    "esModuleInterop": true,
-  }
-}
