+++ syntax/grammar.d.ts
@@ -1,0 +1,10 @@
+import { Resource } from "../resolver/src/ast";
+interface Result<T, E> {
+  fold(s: (value: T) => T, f: (err: E) => never): T;
+}
+
+interface Parser {
+  run(input: string): Result<Resource, string>;
+}
+
+export declare let Resource: Parser;
