+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "target": "es2018",
+        "module": "commonjs",
+        "strict": true,
+        "esModuleInterop": true,
+        "allowJs": true,
+        "outDir": "./build"
+    },
+    "include": ["./syntax/**/*", "./resolver/**/*", "./bin/**/*", "./lib/**/*"]
+}
